app/auth/jwt_handler.py
# app/auth/jwt_handler.py
from datetime import datetime, timedelta
from fastapi import HTTPException, status
from jose import jwt, JWTError
from app.database.config import get_settings

settings = get_settings()

SECRET_KEY = settings.SECRET_KEY
ALGORITHM = settings.JWT_ALGORITHM
ACCESS_TOKEN_EXPIRE_MINUTES = settings.ACCESS_TOKEN_EXPIRE_MINUTES


def create_access_token(payload: dict) -> str:
    """
    Создаёт JWT токен со стандартным exp.
    payload должен содержать:
    {
        "user_id": int,
        "email": str,
        "is_admin": bool
    }
    """
    to_encode = payload.copy()
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode["exp"] = expire

    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)


def decode_access_token(token: str) -> dict:
    """
    Проверяет токен и возвращает payload.
    """
    try:
        return jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
    except JWTError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Токен недействителен или истек",
        )
